Step timings are preserved. Placing the foot a little closer to the edge when stepping up.
By trying to keep the step timings as close as possible to the planned one, it is made sure that the planned trajectories are still valid.
It is possible to place the foot close to the edge. By trial and error, it seems that placing the foot too close to the edge does not help in having shin clearance.